Update imports in moved test files
The unit tests that were moved in the previous commit often imported
utilities from a module that used to be located in the main package
`aiida.backends.tests.utils`. This module has also been moved to the
`tests` top level module.

These utilities should really in most cases become pytest fixtures but
that is left for a later time.
