+++ pytest.ini
@@ -1,5 +1,5 @@
 [pytest]
-testpaths = aiida
+testpaths = tests
 filterwarnings =
     ignore::DeprecationWarning:babel:
     ignore::DeprecationWarning:django:
