+++ tests/cmdline/commands/test_config.py
@@ -12,9 +12,10 @@
 from click.testing import CliRunner
 
 from aiida.backends.testbase import AiidaTestCase
-from aiida.backends.tests.utils.configuration import with_temporary_config_instance
 from aiida.cmdline.commands import cmd_verdi
 from aiida.manage.configuration import get_config
+
+from tests.utils.configuration import with_temporary_config_instance
 
 
 class TestVerdiConfig(AiidaTestCase):
