+++ tests/cmdline/commands/test_verdi.py
@@ -11,9 +11,10 @@
 from click.testing import CliRunner
 
 from aiida import get_version
-from aiida.backends.tests.utils.configuration import with_temporary_config_instance
 from aiida.backends.testbase import AiidaTestCase
 from aiida.cmdline.commands import cmd_verdi
+
+from tests.utils.configuration import with_temporary_config_instance
 
 
 class TestVerdi(AiidaTestCase):
