+++ tests/engine/test_futures.py
@@ -13,9 +13,10 @@ import datetime
 from tornado import gen
 
 from aiida.backends.testbase import AiidaTestCase
-from aiida.backends.tests.utils import processes as test_processes
 from aiida.engine import processes, run
 from aiida.manage.manager import get_manager
+
+from tests.utils import processes as test_processes
 
 
 class TestWf(AiidaTestCase):
