+++ tests/engine/test_process.py
@@ -16,12 +16,13 @@ from plumpy.utils import AttributesFrozendict
 
 from aiida import orm
 from aiida.backends.testbase import AiidaTestCase
-from aiida.backends.tests.utils import processes as test_processes
 from aiida.common.lang import override
 from aiida.engine import ExitCode, ExitCodesNamespace, Process, run, run_get_pk, run_get_node
 from aiida.engine.processes.ports import PortNamespace
 from aiida.manage.caching import enable_caching
 from aiida.plugins import CalculationFactory
+
+from tests.utils import processes as test_processes
 
 
 class NameSpacedProcess(Process):
