+++ tests/manage/configuration/test_config.py
@@ -14,11 +14,12 @@ import shutil
 import tempfile
 
 from aiida.backends.testbase import AiidaTestCase
-from aiida.backends.tests.utils.configuration import create_mock_profile
 from aiida.common import exceptions, json
 from aiida.manage.configuration import Config, Profile, settings
 from aiida.manage.configuration.migrations import CURRENT_CONFIG_VERSION, OLDEST_COMPATIBLE_CONFIG_VERSION
 from aiida.manage.configuration.options import get_option
+
+from tests.utils.configuration import create_mock_profile
 
 
 class TestConfigDirectory(AiidaTestCase):
