+++ tests/manage/configuration/test_profile.py
@@ -13,8 +13,9 @@ import os
 import uuid
 
 from aiida.backends.testbase import AiidaTestCase
-from aiida.backends.tests.utils.configuration import create_mock_profile
 from aiida.manage.configuration import Profile
+
+from tests.utils.configuration import create_mock_profile
 
 
 class TestProfile(AiidaTestCase):
