+++ tests/tools/importexport/migration/test_migration.py
@@ -13,11 +13,12 @@ import os
 
 from aiida import orm
 from aiida.backends.testbase import AiidaTestCase
-from aiida.backends.tests.utils.archives import get_archive_file, get_json_files, migrate_archive
-from aiida.backends.tests.utils.configuration import with_temp_dir
 from aiida.tools.importexport import import_data, EXPORT_VERSION as newest_version
 from aiida.tools.importexport.migration import migrate_recursively, verify_metadata_version
 from aiida.common.utils import Capturing
+
+from tests.utils.archives import get_archive_file, get_json_files, migrate_archive
+from tests.utils.configuration import with_temp_dir
 
 
 class TestExportFileMigration(AiidaTestCase):
