+++ tests/tools/importexport/migration/test_v04_to_v05.py
@@ -13,13 +13,14 @@ import tarfile
 import zipfile
 
 from aiida.backends.testbase import AiidaTestCase
-from aiida.backends.tests.utils.archives import get_archive_file, get_json_files
 from aiida.common.exceptions import NotExistent
 from aiida.common.folders import SandboxFolder
 from aiida.common.json import load as jsonload
 from aiida.tools.importexport.common.archive import extract_tar, extract_zip
 from aiida.tools.importexport.migration.utils import verify_metadata_version
 from aiida.tools.importexport.migration.v04_to_v05 import migrate_v4_to_v5
+
+from tests.utils.archives import get_archive_file, get_json_files
 
 
 class TestMigrateV04toV05(AiidaTestCase):
