+++ tests/tools/importexport/orm/test_links.py
@@ -14,14 +14,15 @@ import tarfile
 
 from aiida import orm
 from aiida.backends.testbase import AiidaTestCase
-from aiida.backends.tests.tools.importexport.utils import get_all_node_links
-from aiida.backends.tests.utils.configuration import with_temp_dir
 from aiida.common import json
 from aiida.common.folders import SandboxFolder
 from aiida.common.links import LinkType
 from aiida.common.utils import get_new_uuid
 from aiida.tools.importexport import import_data, export
 from aiida.tools.importexport.common.exceptions import DanglingLinkError
+
+from tests.utils.configuration import with_temp_dir
+from tests.tools.importexport.utils import get_all_node_links
 
 
 class TestLinks(AiidaTestCase):
