+++ tests/tools/importexport/test_complex.py
@@ -14,9 +14,10 @@ import os
 
 from aiida import orm
 from aiida.backends.testbase import AiidaTestCase
-from aiida.backends.tests.utils.configuration import with_temp_dir
 from aiida.common.links import LinkType
 from aiida.tools.importexport import import_data, export
+
+from tests.utils.configuration import with_temp_dir
 
 
 class TestComplex(AiidaTestCase):
